Revert: Move extension update to the important notifications panel
This reverts commit 11052a090938599bd65d5afffc743c106eaf235f.
