+++ packages/next-common/components/globalNotification/notification.jsx
@@ -2,14 +2,8 @@ import { useCookieValue } from "next-common/utils/hooks/useCookieValue";
 import { GreyPanel } from "next-common/components/styled/containers/greyPanel";
 import { SystemClose } from "@osn/icons/subsquare";
 import { isNil } from "lodash-es";
-import { cn } from "next-common/utils";
 
-export default function CommonNotification({
-  cacheKey,
-  children,
-  expires = 15,
-  className = "",
-}) {
+export default function Notification({ cacheKey, children, expires = 15 }) {
   const [visible, setVisible] = useCookieValue(cacheKey, true);
 
   if (!visible) {
@@ -17,15 +11,10 @@ export default function CommonNotification({
   }
 
   return (
-    <div className={cn("w-full bg-theme500 flex justify-center", className)}>
+    <div className="w-full bg-theme500 flex justify-center">
       <div className="px-6 py-2.5 mx-auto max-w-[1200px] max-sm:px-0 flex-1">
         <div className="w-full px-6">
-          <GreyPanel
-            className={cn(
-              "flex justify-between items-start flex-row bg-theme500 text14Medium rounded-none text-textPrimaryContrast",
-              className,
-            )}
-          >
+          <GreyPanel className="flex justify-between items-start flex-row bg-theme500 text14Medium rounded-none text-textPrimaryContrast">
             <div className="flex-1 justify-center items-center text-center">
               {children}
             </div>
